% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixEQTLwrapper.R
\name{load_sliced_file}
\alias{load_sliced_file}
\title{Load a SlicedData object from file}
\usage{
load_sliced_file(file, sliceSize = 2000, first_col_is_rownames = TRUE)
}
\description{
Load a SlicedData object from file
}
