% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixEQTLwrapper.R
\name{matrixEQTLwrapper}
\alias{matrixEQTLwrapper}
\title{Run MatrixEQTL}
\usage{
matrixEQTLwrapper(
  feature_locations_path,
  feature_data_path,
  snpFilePath,
  covFilePath,
  snpLocPath,
  group_name,
  resultsDir = getwd(),
  cisDist = 1e+06,
  pvOutputThreshold = 1e-05,
  pvOutputThresholdCis = 1e-04,
  useModel = "linear",
  minPvByGeneSnp = TRUE,
  noFDRsaveMemory = FALSE,
  prefix = NULL,
  pvalueHist = NULL,
  verbose = TRUE,
  dry_run = TRUE,
  colNames_convention
)
}
\description{
Run MatrixEQTL
}
