% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nearest_gene_helpers.R
\name{nearest_gene_from_gtf}
\alias{nearest_gene_from_gtf}
\title{Nearest gene (by TSS) from a GTF for given chromosome/position(s)
(patched to safely harmonize seqlevels)}
\usage{
nearest_gene_from_gtf(
  gtf_file,
  chr,
  pos,
  output = c("gene", "distance", "both"),
  tss_gr = NULL
)
}
\description{
Nearest gene (by TSS) from a GTF for given chromosome/position(s)
(patched to safely harmonize seqlevels)
}
