% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixEQTLwrapper.R
\name{validate_sample_overlap}
\alias{validate_sample_overlap}
\title{Validate and reorder sample columns if needed}
\usage{
validate_sample_overlap(feature_data, snps, covs, group_name, verbose = TRUE)
}
\description{
Validate and reorder sample columns if needed
}
